refactoring, cache some of configuration
