+++ pom.xml
@@ -65,6 +65,7 @@
     <surefire.version>3.1.0</surefire.version>
     <jackson.version>2.14.2</jackson.version>
     <gson.version>2.8.9</gson.version>
+    <guava.version>32.1.2-jre</guava.version>
     <commons-compress.version>1.21</commons-compress.version>
     <awaitility.version>4.0.3</awaitility.version>
     <maven.antrun.plugin.version>3.1.0</maven.antrun.plugin.version>
@@ -106,6 +107,11 @@
         <groupId>org.slf4j</groupId>
         <artifactId>slf4j-simple</artifactId>
         <version>${slf4j.version}</version>
+      </dependency>
+      <dependency>
+        <groupId>com.google.guava</groupId>
+        <artifactId>guava</artifactId>
+        <version>${guava.version}</version>
       </dependency>
       <dependency>
         <groupId>org.junit.jupiter</groupId>
